+++ spec/controllers/criteria_controller_spec.rb
@@ -3,6 +3,7 @@ describe CriteriaController do
   let(:assignment) { create :assignment }
   let(:grouping) { create :grouping, assignment: assignment }
   let(:submission) { create :submission, grouping: grouping }
+
   describe 'Using Flexible Criteria' do
     let(:flexible_criterion) do
       create(:flexible_criterion,
@@ -244,7 +245,7 @@ describe CriteriaController do
       context '#create' do
         context 'with save error' do
           before(:each) do
-            expect_any_instance_of(FlexibleCriterion).to receive(:update).and_return(false)
+            expect_any_instance_of(FlexibleCriterion).to receive(:save).and_return(false)
             expect_any_instance_of(FlexibleCriterion).to receive(:errors).and_return(ActiveModel::Errors.new(self))
             post_as admin,
                     :create,
@@ -601,11 +602,11 @@ describe CriteriaController do
       context '#create' do
         context 'with save error' do
           before(:each) do
-            expect_any_instance_of(RubricCriterion).to receive(:update).and_return(false)
+            expect_any_instance_of(RubricCriterion).to receive(:save).and_return(false)
             expect_any_instance_of(RubricCriterion).to receive(:errors).and_return(ActiveModel::Errors.new(self))
             post_as admin,
                     :create,
-                    params: { assignment_id: assignment.id, rubric_criterion: { name: 'first', max_mark: 10 },
+                    params: { assignment_id: assignment.id, max_mark_prompt: 10,
                               new_criterion_prompt: 'first', criterion_type: 'RubricCriterion' },
                     format: :js
           end
@@ -623,7 +624,7 @@ describe CriteriaController do
           before(:each) do
             post_as admin,
                     :create,
-                    params: { assignment_id: assignment.id, rubric_criterion: { name: 'first' },
+                    params: { assignment_id: assignment.id,
                               new_criterion_prompt: 'first', criterion_type: 'RubricCriterion', max_mark_prompt: 10 },
                     format: :js
           end
@@ -792,36 +793,23 @@ describe CriteriaController do
 
       it 'creates rubric criteria with properly formatted entries' do
         post_as admin, :upload, params: { assignment_id: assignment.id, upload_file: mixed_file }
-
         expect(assignment.get_criteria(:all, :rubric).pluck(:name)).to contain_exactly('cr30', 'cr90')
-
         cr1 = assignment.get_criteria(:all, :rubric).find_by(name: 'cr30')
+        expect(cr1.levels.size).to eq(5)
         expect(cr1.max_mark).to eq(5.0)
-        expect(cr1.level_0_name).to eq('What?')
-        expect(cr1.level_0_description).to eq('Fail')
-        expect(cr1.level_1_name).to eq('Hmm')
-        expect(cr1.level_1_description).to eq('Almost fail')
-        expect(cr1.level_2_name).to eq('Average')
-        expect(cr1.level_2_description).to eq('Not bad')
-        expect(cr1.level_3_name).to eq('Good')
-        expect(cr1.level_3_description).to eq('Alright')
-        expect(cr1.level_4_name).to eq('Excellent')
-        expect(cr1.level_4_description).to eq('Impressive')
         expect(cr1.ta_visible).to be false
         expect(cr1.peer_visible).to be true
+        # Since there are only 5 levels in this rubric criterion, if each of the following queries return an entity,
+        # then this rubric criterion is properly sat up.
+        expect(cr1.levels.find_by(name: 'Beginner', description: 'Fail', mark: 0)).not_to be_nil
+        expect(cr1.levels.find_by(name: 'Hmm', description: 'Almost fail', mark: 1)).not_to be_nil
+        expect(cr1.levels.find_by(name: 'Average', description: 'Not bad', mark: 2)).not_to be_nil
+        expect(cr1.levels.find_by(name: 'Good', description: 'Alright', mark: 3)).not_to be_nil
+        expect(cr1.levels.find_by(name: 'Excellent', description: 'Impressive', mark: 5)).not_to be_nil
 
         cr2 = assignment.get_criteria(:all, :rubric).find_by(name: 'cr90')
         expect(cr2.max_mark).to eq(4.6)
-        expect(cr2.level_0_name).to be_nil
-        expect(cr2.level_0_description).to be_nil
-        expect(cr2.level_1_name).to be_nil
-        expect(cr2.level_1_description).to be_nil
-        expect(cr2.level_2_name).to be_nil
-        expect(cr2.level_2_description).to be_nil
-        expect(cr2.level_3_name).to be_nil
-        expect(cr2.level_3_description).to be_nil
-        expect(cr2.level_4_name).to be_nil
-        expect(cr2.level_4_description).to be_nil
+        expect(cr2.levels.size).to eq(5)
         expect(cr2.ta_visible).to be true
         expect(cr2.peer_visible).to be false
       end
@@ -891,7 +879,6 @@ describe CriteriaController do
 
       it 'creates criteria with rounded (up to first digit after decimal point) maximum mark' do
         post_as admin, :upload, params: { assignment_id: assignment.id, upload_file: round_max_mark_file }
-
         expect(assignment.get_criteria(:all, :rubric).first.name).to eq('cr90')
 
         expect(assignment.get_criteria(:all, :rubric).first.max_mark).to eq(4.6)
@@ -921,11 +908,26 @@ describe CriteriaController do
 
         criteria = assignment.get_criteria(:all, :rubric).first
         expect(criteria.name).to eq('Quality of Writing')
-        expect(criteria.level_4_name).to be_nil
-        expect(criteria.level_4_description).to be_nil
-        expect(flash[:success].size).to eq(1)
+        expect(criteria.levels.size).to eq 6
+        criteria.levels.each do |level|
+          expect(level.valid?).to eq true
+        end
+        expect(criteria.levels[0].name).to eq('Beginner')
+        expect(criteria.levels[0].description).to eq('The essay is very poorly organized'\
+                                                       ' structure and gives no new information.')
+        expect(criteria.levels[0].mark).to eq(10.0)
 
-        expect(assignment.get_criteria.map(&:name)).not_to include('Level 5')
+        expect(criteria.levels[1].name).to eq('Capable')
+        expect(criteria.levels[1].description).to eq('The essay is poorly organized but gives new information.')
+        expect(criteria.levels[1].mark).to eq(14.0)
+
+        expect(criteria.levels[2].name).to eq('Accomplished')
+        expect(criteria.levels[2].description).to eq('The essay is well-structure and conveys new information clearly.')
+        expect(criteria.levels[2].mark).to eq(18.0)
+
+        expect(criteria.levels[3].name).to eq('Level 3')
+        expect(criteria.levels[3].description).to eq('Level 3 description in one line.')
+        expect(criteria.levels[3].mark).to eq(22.0)
 
         pending('We should report there is an invalid key in the file')
         expect(flash[:error]).not_to be_nil
@@ -942,16 +944,15 @@ describe CriteriaController do
 
         it 'sends the correct information' do
           post_as admin, :upload, params: { assignment_id: assignment.id, upload_file: test_upload_download_file }
-
           get :download, params: { assignment_id: assignment.id }
 
-          expect(YAML.safe_load(response.body)).to eq(YAML.safe_load(expected_download.read))
+          expect(YAML.safe_load(response.body, [Symbol], symbolize_names: true))
+            .to eq(YAML.safe_load(expected_download.read, symbolize_names: true))
         end
       end
     end
   end
 
-  let(:assignment) { FactoryBot.create(:assignment) }
   context '#upload', pending: true do # Until criteria tables merged together, can't use Criterion.count
     include_examples 'a controller supporting upload' do
       let(:params) { { assignment_id: assignment.id } }