+++ spec/factories/rubric_criteria.rb
@@ -5,16 +5,11 @@ FactoryBot.define do
     max_mark { 4.0 }
     ta_visible { true }
     peer_visible { false }
-    level_0_name { 'Poor' }
-    level_0_description { 'This criterion was not satisifed whatsoever.' }
-    level_1_name { 'Satisfactory' }
-    level_1_description { 'This criterion was satisfied.' }
-    level_2_name { 'Good' }
-    level_2_description { 'This criterion was satisfied well.' }
-    level_3_name { 'Great' }
-    level_3_description { 'This criterion was satisfied very well.' }
-    level_4_name { 'Excellent' }
-    level_4_description { 'This criterion was satisfied exceptionally well.' }
     sequence(:position)
+    after(:build) do |criterion|
+      5.times.each do |i|
+        criterion.levels << build(:level, rubric_criterion: criterion, mark: criterion.max_mark * i / 4)
+      end
+    end
   end
 end